+++ components/SignupProducts/GenericProduct.tsx
@@ -1,43 +1,26 @@
 import Input from "@/components/Form/Input";
 import Dropdown from "@/components/Form/Dropdown";
 import { useTranslation } from "react-i18next";
+import Location from "@/components/SignupProducts/common/Location";
+import Quantity from "@/components/SignupProducts/common/Quantity";
+import Product from "@/components/SignupProducts/common/Product";
 
 const GenericProduct = () => {
   const { t } = useTranslation();
 
   return (
     <div>
-      <Input
-        name="products_food_name"
-        label={t('signup.products.food.name')}
-        labelPosition="horizontal"
-      />
-      <div className="flex">
-        <Input
-          name="products_food_qty"
-          type="number"
-          label={t('signup.products.food.qty')}
-          labelPosition="horizontal"
-        />
-        <Input
-          name="products_food_packaging"
-          label={t('signup.products.food.packaging')}
-          labelPosition="horizontal"
-        />
-      </div>
+      <Product />
+
+      <Quantity />
+
       <Input
         type="date"
         name="product_food_expire_date"
         label={t('signup.products.food.expire_date')}
       />
-      <div className="flex">
-        <Dropdown name="product_food_county">
-          <option>Cluj</option>
-        </Dropdown>
-        <Dropdown name="product_food_city">
-          <option>Cluj Napoca</option>
-        </Dropdown>
-      </div>
+
+      <Location />
     </div>
   )
 }
